+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2022-05-21 v0.0.2
+- Added UserWeb examples
+- Moved cometsdk to its own lib module
+- Added fix for some marshalling issues
+
 ## 2022-07-28 v0.0.1
 - Based on 22.6.8
 - Initial release